requirements.txt
Django>3.2,<3.5
djangorestframework>3.12,<3.15
django-cors-headers>=3.7,<4
django-filter>=2.4,<3